Fix #3176: Rotate the vertices to match the marker
This applies the rotation offset value to the list of vertices
before we return that list to the competitors code. It would be
nice if libkoki itself did this, but since there are examples
which also do this (see draw_marker in realtime_gl.c), care would
need to be taken adjusting that as it would be a breaking change.

For similar reasons I don't think that this change should ship
during the SR year.

Change-Id: I3b0bd9141b8a34f38ac0d19a195310c7cf3463a6